tools: test_gw_repeater: start 1st repeater without vendor-specific
Since we want to run our tests as close as possible to the standard, run
the repeaters as non-intel by passing the -n option to run.sh. This has
to be done after a -- because the -n is to be interpreted by the
underlying script, not by run.sh itself. For symmetry, also add a -- to
the gateway start.

Only the first repeater is configured like that, so we still test both
variants.

The gateway has to remain an Intel agent because the controller is
blocked until an Intel agent joins it.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>