gitlab-ci.yml: do build step directly in builder image
Instead of pulling the docker images in the build step, use a normal
runner and build directly in the prplmesh-builder image pulled from
dockerhub. Note that pulling the runner image still needs to be done in
the run-tests step.

Note that we now use the clang-format.sh script from the top-level
directory instead of the one from tools/docker. The latter one just
calls the one in the top-level directory inside the prplmesh-builder
container.

The build results will still be in the same location and passed on to
the next stage.

Since the build step now uses the normal docker-based build process, it
is not needed to specify tags for it.

The before and after scripts are only needed for the run-tests stage, so
move them there.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>