tools: docker: add an option to push images
It's more convenient to be able to automatically push the images to a
registry after they have been built (so that they can be used by
others).

Add a new function: push, to push a tag and check if it succeeded.

Add a new option: -p or --push, to add a registry to push to.

After generating the images if a push registry was provided, tag the
image with the given registry, and use the new push function to push
it.

Also fix the long options requiring an argument (they missed a colon).

While we're at it, add the missing quotes around "$@", and initialize
$TAG to an empty string.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>