tlvf: add resizeTlvs for inner TLV lists
A main TLV list may contain an inner TLV list, which is used to create /
parse a specific TLV payload.
Examples for such are the WSC TLV which contains a list of WSC
attributes, and the vendor specific TLV which contains arbitrary data
so can also be managed via an inner TLV list.

TLVs which contain inner TLV lists contain an unknown length list of
uint8_t - the payload, which is passed to the inner TLVlist constructor
as the buffer and buffer length.

When the payload is allocated, it may or may not contain variable length
data (such as variable length lists, strings, unknown length lists,
etc.), so we normally pre-allocate the payload to the maximum allowed
size. Later, on finalize(), we need to trim the tailroom of any unused
buffer space since some TLVs such as the WSC TLV does not allow to have
extranous data at the end.

This commit adds the resizeTlvs() private function to TlvList class. It
is called __only__ from the finalize() call, calculates the tailroom
of each inner TLV list, and trims the tailroom by moving the rest of the
buffer to the current TLVlist's buffer pointer.

Then it updates the tailroom of the parent TLV so the length calculation
coming afterwards is correct.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>