tlvf: tlvlist add inner tlvlist
Preperative commit.
Add an unordered map of TLVList to the TLVlist base class which will be
used for creating/parsing of TLV internal data.

Each outer TLV will have an unknown length list for the inner TLV list
data:

outer:
	_type: class
	_is_tlv_class: true
	type: eTlvType
	length: uint16_t
	payload:
		_type: uint8_t
		_length: []

The payload is used as the buffer for the inner tlv list.
This will be used for implementing the following:
- Intel Vendor specific TLVs data - action header + action
- WSC TLVs AttrList

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>