tlvf: add tlvWsc.yaml
Preparative commit.

The WSC TLVs are regular TLVs which contain a list of WSC Attributes,
where each attribute is of a TLV structure where both length and type
are 2 bytes long.

The current implementation defines 2 yaml files for 2 types of WSC TLVs
- WSC M2 and WSC M1. Each of those yaml files contain the list of
attributes (defined in WSC_Attributes.yaml) in strict ordering, meaning
that if we get a WSC M2 or M1 TLV with different attributes order then
we will fail on parsing.

This is by itself wrong, and the reason that prplMesh cannot currently
pass autoconfiguration with most controller when acting as agent.

Previous PR (#572) added support for TlvList class, which can be used to
describe a list of TLV like structures inside an outer TLV.
The PR did it for vendor specific Intel messages by adding
beerocks_header which derives from TlvList.

The same solution can now be used to fix the ordering issue with the WSC
TLVs - by defining a new WSC TLV yaml with an unknown length list of
uint8_t which will be used as the buffer for the inner TlvList.

This commit adds WscTlv.yaml which provides just that. Followup commits
will use this to acheive the final goal of replacing tlvWscM1.yaml and
tlvWscM2.yaml with the wscTlv.yaml and a derived class of TlvList which
will provide parsing and create APIs for both M2 and M1 WSC types.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>