.gitlab-ci.yml: build for the Turris Omnia
Add a new hidden job: build-for-openwrt. It is defined to make it easy
to build for different targets. It has to be a hidden job because it's
not supposed to be run as-is.

Also add a new job extending .build-for-openwrt, to build for the
Turris Omnia.

Add a "needs" for build-in-docker to the run-tests job, so that it
doesn't wait for the OpenWrt build nor use its artifacts.

While we're at it, remove the duplicate "artifacts" for the
run_tests job.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>