tools: openwrt: build.sh: switch to options instead of variables
Using command line options makes it easier to control what the build
script does.

Add command line options to build.sh, and drop the "args" files.

The CI configuration is adapted to use the new options: drop the
target file, add the '-d' option. While we're at it, use '-t' to give
a unique tags to the containers, so that the job can be run several
times at the same time on the same runner.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>