tools: docker: make the openwrt build script stop at errors
When using the build script in CI, it's essential that it either:
- records errors and exits with a corresponding exit code at the end,
or
- fail at the first error.

Otherwise, the CI could report a successful job while in reality it
failed.

Add the '-e' option to the build script, so that it stops with an
error code at the first failing command.

Also add the '-v' option to the last cp command, to be verbose about
what is being copied.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>