controller: client_steering_task: wait for disconnect and connect
The steering task is considered successful when the connected event for
the STA is received. However, we were not waiting for that event, so if
the disconnected event arrives before the connected event, we would
first jump to the FINALIZE state and report that steering has failed.

Add STA_CONNECTED to the events that we're waiting for.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>