tlvf: AttrList move M1 and M2 getters
Preparative commit.
An AttrList should not be bound to M1 and M2 only, so the getter
functions even if they are common to M1 and M2 should be part of the M1
and M2 classes and not the base class.
This commit moves all common getters to the M1 and M2 classes.

Signed-off-by: Tomer Eliyahu <tomereliyahu1@gmail.com>