controller: always inform steering task of disconnect
The disconnect event was only passed to the steering task if the
handoff_flag was false, i.e. if there was no steering ongoing. This will
obviously never work. Move the steering task event out of the handoff
condition.

While we're at it, only push it to the steering task if the steering
task is running - otherwise push_event will print an error message.

Also while we're at it, update the comments to clarify what is going on.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>