common: bcl: network_utils: add elpp wrapper for sMacAddr
We very often need to log a MAC address. Instead of calling
network_utils::mac_to_string all the time, it's much easier to define an
additional operator<< for elpp that handles sMacAddr.

We could do the same for std::ostream, but since we're not using that at
all, there's not reason to add it.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>