agent: slave: use radio MAC as radio identifier
The Intel code has two identifiers for the radio: a uid and a mac.
The Multi-AP specification, on the other hand, only has a Radio (Unique)
Identifier. We need to choose one of them, but that choice was not made
very consistently.

The radio MAC is the most convenient one to use - it is for example what
gets printed in bml_conn_map. So switch the agent to use that instead of
the radio_identifier.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>